CI/CD/opa/cloud-resources/requirements.txt
yamllint>=1.20.0
j2cli[yaml]>=0.3.10
pyaml>=19.12.0
cerberus>=1.3.2
PyYAML==5.3.1
HiYaPyCo==0.4.16
